make rounded background its own component, use it in engage with us
